Fix download_benchmarking_data()

+++ dlclive/benchmark.py
@@ -37,35 +37,34 @@ from dlclive.utils import decode_fourcc
 
 def download_benchmarking_data(
     target_dir=".",
-    url="https://huggingface.co/datasets/mwmathis/DLCspeed_benchmarking/blob/main/Data-DLC-live-benchmark.zip",
+    url="https://huggingface.co/datasets/mwmathis/DLCspeed_benchmarking/resolve/main/Data-DLC-live-benchmark.zip",
 ):
     """
-    Downloads a DeepLabCut-Live benchmarking Data (videos & DLC models).
+    Downloads and extracts DeepLabCut-Live benchmarking data (videos & DLC models).
     """
     import os
     import urllib.request
     from tqdm import tqdm
     import zipfile
 
-    def show_progress(count, block_size, total_size):
-        pbar.update(block_size)
-
-    response = urllib.request.urlopen(url)
-    total_size = int(response.getheader("Content-Length"))
-    pbar = tqdm(unit="B", total=total_size, position=0, desc="Downloading")
-    filename, _ = urllib.request.urlretrieve(url, filename=zip_path, reporthook=show_progress)
-    pbar.close()
-
-    class DownloadProgressBar(tqdm):
-        def update_to(self, b=1, bsize=1, tsize=None):
-            if tsize is not None:
-                self.total = tsize
-            self.update(b * bsize - self.n)
+    # Avoid nested folder issue
+    if os.path.basename(os.path.normpath(target_dir)) == "Data-DLC-live-benchmark":
+        target_dir = os.path.dirname(os.path.normpath(target_dir))
+    os.makedirs(target_dir, exist_ok=True)  # Ensure target directory exists
 
     zip_path = os.path.join(target_dir, "Data-DLC-live-benchmark.zip")
-    print(
-        f"Downloading the benchmarking data from {url} ..."
-    )
+
+    if os.path.exists(zip_path):
+        print(f"{zip_path} already exists. Skipping download.")
+    else:
+        def show_progress(count, block_size, total_size):
+            pbar.update(block_size)
+
+        print(f"Downloading the benchmarking data from {url} ...")
+        pbar = tqdm(unit="B", total=0, position=0, desc="Downloading")
+
+        filename, _ = urllib.request.urlretrieve(url, filename=zip_path, reporthook=show_progress)
+        pbar.close()
 
     print(f"Extracting {zip_path} to {target_dir} ...")
     with zipfile.ZipFile(zip_path, 'r') as zip_ref:
